Bump version of snakemake.
Because of a bug (https://github.com/snakemake/snakemake/issues/1891)
we need to either pin tabulate to an older version or bump snakemake
to a newer version. We opt for the second solution, as there is no
reason to use an outdated version of snakemake.

Note that https://github.com/astanin/python-tabulate/pull/232
would also fix the issue, but this isn't released yet in tabulate.